Uncomment groovy scripts
