+++ buildSrc/src/main/groovy/js/js.gradle
@@ -54,120 +54,120 @@ apply from: "$rootDir" + io.spine.internal.gradle.Scripts.commonPath + "js/npm-p
 /**
  * Cleans old module dependencies and build outputs.
  */
-//task deleteCompiled {
-//    group = JAVA_SCRIPT_TASK_GROUP
-//    description = 'Cleans old module dependencies and build outputs.'
-//
-//    clean.dependsOn deleteCompiled
-//
-//    doLast {
-//        delete genProtoMain
-//        delete genProtoTest
-//        delete coverageJs.outputs
-//    }
-//}
+task deleteCompiled {
+    group = JAVA_SCRIPT_TASK_GROUP
+    description = 'Cleans old module dependencies and build outputs.'
+
+    clean.dependsOn deleteCompiled
+
+    doLast {
+        delete genProtoMain
+        delete genProtoTest
+        delete coverageJs.outputs
+    }
+}
 
 /**
  * Customizes the task already defined in `config` module by running Webpack build.
  */
-//buildJs {
-//
-//    outputs.dir "$projectDir/dist"
-//
-//    doLast {
-//        npm 'run', 'build'
-//        npm 'run', 'build-dev'
-//    }
-//}
+buildJs {
+
+    outputs.dir "$projectDir/dist"
+
+    doLast {
+        npm 'run', 'build'
+        npm 'run', 'build-dev'
+    }
+}
 
 /**
  * Customizes the task already defined in `config` module by running
  * the JavaScript tests.
  */
-//testJs {
-//
-//    doLast {
-//        npm 'run', 'test'
-//    }
-//}
+testJs {
+
+    doLast {
+        npm 'run', 'test'
+    }
+}
 
 /**
  * Copies bundled JS sources to the temporary NPM publication directory.
  */
-//task copyBundledJs(type: Copy) {
-//    group = JAVA_SCRIPT_TASK_GROUP
-//    description = 'Copies assembled JavaScript sources to the NPM publication directory.'
-//
-//    from buildJs.outputs
-//    into "$publicationDirectory/dist"
-//
-//    dependsOn buildJs
-//}
+task copyBundledJs(type: Copy) {
+    group = JAVA_SCRIPT_TASK_GROUP
+    description = 'Copies assembled JavaScript sources to the NPM publication directory.'
+
+    from buildJs.outputs
+    into "$publicationDirectory/dist"
+
+    dependsOn buildJs
+}
 
 /**
  * Transpiles JS sources before publishing them to NPM.
  *
  * Puts the resulting files to the temporary NPM publication directory.
  */
-//task transpileSources {
-//    group = JAVA_SCRIPT_TASK_GROUP
-//    description = "Transpiles sources before publishing."
-//
-//    doLast {
-//        npm 'run', 'transpile-before-publish'
-//    }
-//}
+task transpileSources {
+    group = JAVA_SCRIPT_TASK_GROUP
+    description = "Transpiles sources before publishing."
+
+    doLast {
+        npm 'run', 'transpile-before-publish'
+    }
+}
 
 /**
  * Defines files to copy by the task.
  */
-//prepareJsPublication {
-//
-//    doLast {
-//        copy {
-//            from (projectDir) {
-//                include 'package.json'
-//                include '.npmrc'
-//            }
-//            into publicationDirectory
-//        }
-//    }
-//
-//    //TODO:2019-02-05:dmytro.grankin: temporarily don't publish a bundle, see https://github.com/SpineEventEngine/web/issues/61
-//    //dependsOn copyBundledJs
-//    dependsOn transpileSources
-//}
+prepareJsPublication {
+
+    doLast {
+        copy {
+            from (projectDir) {
+                include 'package.json'
+                include '.npmrc'
+            }
+            into publicationDirectory
+        }
+    }
+
+    //TODO:2019-02-05:dmytro.grankin: temporarily don't publish a bundle, see https://github.com/SpineEventEngine/web/issues/61
+    //dependsOn copyBundledJs
+    dependsOn transpileSources
+}
 
 /**
  * Runs the JavaScript tests and collects the code coverage.
  */
-//task coverageJs {
-//    group = JAVA_SCRIPT_TASK_GROUP
-//    description = 'Runs the JS tests and collects the code coverage info.'
-//
-//    outputs.dir nycOutputDir
-//
-//    final def runsOnWindows = org.gradle.internal.os.OperatingSystem.current().isWindows()
-//    final def coverageScript = runsOnWindows ? 'coverage:win' : 'coverage:unix'
-//
-//    doLast {
-//        npm 'run', coverageScript
-//    }
-//
-//    dependsOn buildJs
-//
-//    rootProject.check.dependsOn coverageJs
-//}
+task coverageJs {
+    group = JAVA_SCRIPT_TASK_GROUP
+    description = 'Runs the JS tests and collects the code coverage info.'
+
+    outputs.dir nycOutputDir
+
+    final def runsOnWindows = org.gradle.internal.os.OperatingSystem.current().isWindows()
+    final def coverageScript = runsOnWindows ? 'coverage:win' : 'coverage:unix'
+
+    doLast {
+        npm 'run', coverageScript
+    }
+
+    dependsOn buildJs
+
+    rootProject.check.dependsOn coverageJs
+}
 
 /**
  * Generates the report on NPM dependencies and their licenses.
  */
-//task npmLicenseReport {
-//
-//    doLast {
-//        npm 'run', 'license-report'
-//    }
-//}
+task npmLicenseReport {
+
+    doLast {
+        npm 'run', 'license-report'
+    }
+}
 
 /**
  * Runs NPM license report straight after the Gradle license report.
@@ -175,42 +175,42 @@ apply from: "$rootDir" + io.spine.internal.gradle.Scripts.commonPath + "js/npm-p
 //TODO:2021-09-22:alexander.yevsyukov: Resolve this dependency.
 //generateLicenseReport.finalizedBy npmLicenseReport
 
-//apply plugin: 'io.spine.mc-js'
-//
-//protoJs {
-//    generatedMainDir = genProtoMain
-//    generatedTestDir = genProtoTest
-//
-//    generateParsersTask().dependsOn compileProtoToJs
-//    buildJs.dependsOn generateParsersTask()
-//    testJs.dependsOn buildJs
-//}
+apply plugin: 'io.spine.mc-js'
+
+protoJs {
+    generatedMainDir = genProtoMain
+    generatedTestDir = genProtoTest
+
+    generateParsersTask().dependsOn compileProtoToJs
+    buildJs.dependsOn generateParsersTask()
+    testJs.dependsOn buildJs
+}
 
 
-//protobuf {
-//    generatedFilesBaseDir = genProtoBaseDir
-//    protoc {
-//        artifact = io.spine.internal.dependency.Protobuf.compiler
-//    }
-//    generateProtoTasks {
-//        all().each { final task ->
-//            task.builtins {
-//                // Do not use java builtin output in this project.
-//                remove java
-//
-//                // For information on JavaScript code generation please see
-//                // https://github.com/google/protobuf/blob/master/js/README.md
-//                js {
-//                    option "import_style=commonjs"
-//                    outputSubDir = genProtoSubDir
-//                }
-//
-//                task.generateDescriptorSet = true
-//                final def testClassifier = task.sourceSet.name == "test" ? "_test" : ""
-//                final def descriptorName = "${project.group}_${project.name}_${project.version}${testClassifier}.desc"
-//                task.descriptorSetOptions.path = "${projectDir}/build/descriptors/${task.sourceSet.name}/${descriptorName}"
-//            }
-//            compileProtoToJs.dependsOn task
-//        }
-//    }
-//}
+protobuf {
+    generatedFilesBaseDir = genProtoBaseDir
+    protoc {
+        artifact = io.spine.internal.dependency.Protobuf.compiler
+    }
+    generateProtoTasks {
+        all().each { final task ->
+            task.builtins {
+                // Do not use java builtin output in this project.
+                remove java
+
+                // For information on JavaScript code generation please see
+                // https://github.com/google/protobuf/blob/master/js/README.md
+                js {
+                    option "import_style=commonjs"
+                    outputSubDir = genProtoSubDir
+                }
+
+                task.generateDescriptorSet = true
+                final def testClassifier = task.sourceSet.name == "test" ? "_test" : ""
+                final def descriptorName = "${project.group}_${project.name}_${project.version}${testClassifier}.desc"
+                task.descriptorSetOptions.path = "${projectDir}/build/descriptors/${task.sourceSet.name}/${descriptorName}"
+            }
+            compileProtoToJs.dependsOn task
+        }
+    }
+}
